Change temperature to be a small positive integer too

+++ coffee/models.py
@@ -2,7 +2,7 @@ from django.db import models
 
 
 class Coffee(models.Model):
-    temperature = models.PositiveIntegerField()
+    temperature = models.PositiveSmallIntegerField()
     cups = models.PositiveSmallIntegerField()
     amount = models.PositiveSmallIntegerField()
